Let imenu distinguish functions by arity
Make the function names used by M-x imenu contain the function arity
as well as the name.  This makes it possible to jump to all functions
of the same name, as opposed to only the one that is defined first.
For example, in the following module there is now a way to jump
directly to foo/1:

-module(foo).

foo() ->
    ok.

bar() ->
    error.

foo(A) ->
    {ok, A}.