Remove tid() from the predefined builtin types.
Change erl_lint not to recognize this type as builtin and
add a new erl_lint.beam version in bootstrap.

Add an -opaque type declaration for this type in ets.erl
and also declare this as an exported type. Use this type
in file debugger/src/dbg_iload.erl in a spec.
While at it, also clean up this later file a bit.