Fix so that ct_run converts relative diretories in the code path to absolute
Directories added with run_test -pa or -pz in the pre erl_args part of the command line will be converted from relative to absolute, this to avoid confusion when Common Test switches working directory during the test run.

Also fixed in this commit: Problem with timeouts during init_tc or end_tc (in the Test Server framework).