beam_asm: Simplify string table generation for beams
The code for generating the string table (which is now
only used for bit syntax matching) in a BEAM file is quite
complicated and potentially expensive when compiling modules
with many thousands of clauses doing bit syntax matching.

Simplify and optimize the code using bit syntax and
binary:match/2 instead of the list operations in the
original code.