Fix building of primary bootstrap when versions have changed
If the versions of kernel or stdlib have changed, attempting
to update the primary bootstrap (using "./otp_build update_primary")
will fail during the rebuild of the start scripts.

The problem is that the app files in bootstrap/lib/*/ebin/*.app
will not be refreshed if they exist, because of missing dependencies.
Ultimately the dependencies should be fixed, but since there
already is a bootstrap_clean target to work around missing
dependencies, fix bootstrap_clean to also remove
bootstrap/lib/*/ebin/*.app.