fix hipe_bifs_alloc_data_2 to avoid "Yikes!" warning
It's been reported that HiPE-enabled Erlang VMs running on BSD
systems sometimes generate messages like

Yikes! erts_alloc() returned misaligned address 0x8016a512c

These originate from hipe_bif0.c:hipe_bifs_alloc_data_2().

A native code module has an associated data area of some
size and alignment.  In the case where the size is zero,
the alignment is irrelevant, but the allocation BIF checks
it anyway.  The warning then triggers on systems where
malloc(0) returns blocks with less alignment than we
(erroneously) expected.

The fix is to simply skip the allocation in this case and
return NULL.  The loader won't actually use the address in
this case so that's safe.  This is also an optimization since
it avoids allocating memory that cannot be used, and it avoids
fragmenting the system heap with useless tiny blocks.

A second problem is that the warning message failed to
identify its origin.  Fixed by prefixing the message by
the BIF's name rather than the silly Yikes! string.

Tested and confirmed to solve the original reporter's problem.