Solaris/x86: Handle floating point exceptions properly in driver threads
On Solaris/x86, the floating point exceptions bits seems to
be propagated to child threads, so if scheduler threads
use floating points exceptions, also driver threads will
have floating points exception enabled.

The problem is that erts_thread_init_float() has not been
called for driver threads, and because of that matherr()
will crash when called.

Fix matherr() so that it will work even if erts_thread_init_float()
has not been called.

This fix was suggested by Mikael Pettersson.