fix crash in hipe_icode_exceptions
This fixes a problem in the HiPE compiler's BEAM-to-Icode
pass where it can crash due to the BEAM compiler's merging
of identical basic blocks with different exception handling
contexts.  This bug has been latent, but is now exposed by
what appears to be more aggressive optimizations in BEAM.

The added comment explains things in more detail.