Change pool module to attempt to attach to nodes that are already running
The pool module prints out an error message and takes no
further action for nodes that are already running. This
patch changes that behavior so that if the return from
slave:start/3 is {already_running, Node} then an attempt
to attach to the node is still made. This makes sense
because the node has been specified by the user in the
.hosts.erlang file indicating a wish for the node to
be part of the pool and a manual attach can be
successfully made after the pool is started.