Make gen_tcp:recv/2 consistent with ssl:recv/2
When the HTTP packet mode has been enabled for a socket,
the ssl and gen_tcp modules have different error indications
when there is an error while parsing the HTTP header:

  ssl:recv(SSLSocket, 0) -> {ok, {http_error, _Str}}

  gen_tcp:recv(Socket, 0) -> {error, {http_error, _Str}}

We have decided to change gen_tcp:recv/2 to behave the same
way as ssl:recv/2. That means that there will be always be
an ok tuple if data could be succefully read from the socket,
and an error tuple if there was a read error at the socket level.