Don't generate multiple tail segments in binary matching
In binary matching, there must only be one "tail segment" (i.e.
a size-less segment of type binary) and it must be last. Thus,
the compiler will reject the following function definition:

  foo(<<A/bits,B/bits>>) -> ok.

But code such as the following:

  [42 || <<_:8/integer, _/bits>> <= Bits]

will internally (in the Core Erlang format) be translated to a
binary matching pattern containing two tail segments. The compiler
happens to generate correct code anyway (later passes will get
rid of the redundant tail segment), but it is ugly and will
confuse tools such as Dialyzer.

Change the transformation of binary generators (in both list and
binary comprehensions) not to generate add a tail segment if there
already is one.