Fix a bug in dialyzer_analysis_callgraph module, occurs when selecting multiple directories and files to analyze from the dialyzer GUI.
This patch fix a bug found in dialyzer GUI, when the user selects multiple files and directories. When dialyzer_analysis_graph tries to expand the files list, it will only take the list of files under the last directory in the list. The reason for that is the function expand_files/3 in module dialyzer_analysis_callgraph. When the function process a directory, it will replace the accumulator with the list of files under that directory.
The fix chosen was to add the files under that directory to the old accumulator to maintain the complete correct list of files to analyze.