dialyzer: Speed up analysis of nested list comprehensions
Nested list comprehensions are translated into strongly connected
funs but dialyzer ignored this. This meant that self-recursive
analysis was normally executed for each fun and the whole fixpoint
took some time to calculate. This patch adds every fun found in
constraint generation to the SCC that is under analysis and then
solves the SCC as a whole.

Possible issues:
- The returned dict contains more entries than usual. This
  triggered a bug in contract checking that is dealt with is
  this patch as well (dialyzer_contracts.erl).
- As it's not easy to tell apart real fun SCCs from simple funs it
  performance might be hampered in the simple funs case.

While doing these changes, also added and strengthened some specs.

In addition, incorporated a modified patch by Ahmed Omar that fixes
a bug in dialyzer_analysis_callgraph module that kept the files of
only one selected directory in dialyzer's GUI.