Remove the last vestiges of the allocating fmove/2 instruction
There was a version of the BEAM loader and emulator that
had two versions of the fmove/2 instruction, one version
that allocated heap space internally and a newer version that
assumed that a previous test_heap/2 instruction had already
allocated the heap space.

Though the allocating fmove/2 instruction is no longer
supported, some vestiges of it still remains.