Added alert in stream cipher case.
Also changed alert to BAD_RECORD_MAC as:

"differentiating between bad_record_mac and decryption_failed alerts
may permit certain attacks against CBC mode as used in TLS
[CBCATT].  It is preferable to uniformly use the bad_record_mac
alert to hide the specific type of the error."

Also cleaned up the code and changed a few other alert reasons in
according to alert descriptions in the TLS RFC 4346. And added function
terminate_alert/3 so that we can differentiate between a crash
in ssl (a bug in our code) and a crash in the application using ssl.