beam_utils: Fix check_liveness/3 for receive loops
Sometimes the beam_bool pass wants to know whether an
y register will be killed by the code that follows and
will do (effectively):

  beam_utils:is_killed({y,Y}, Code, L)

When asked to calculate the liveness for an y register,
beam_utils:is_killed/3 will loop forever if the code
includes a receive loop.

Since this rarely occurs, fix the problem in the simplest
and most conservative way.

Reported-by: Christopher Williams