Fix type specification of the ets:new/2 BIF
The options for the ets:new/2 call has changed recently. There are
read_concurrency hints as well as compressed tables. The hipe/dialyzer
did not take this into account. This patch corrects the problem by
going through the documentation and altering the type specification.

The error fixed is with the dialyzer. When constructing an ETS table
with, e.g., {read_concurrency, true}, the dialyzer will report that the
caller of ets:new/2 will not return.

; First, we update the documentation from the ets man page.
; Second, we reorder the type specification of keypos so they appear
  in the same order as in the documentation.
; Finally, we add the missing read_concurrency and compressed Tweak
  options.