Add an option erlang-icr-indent
The option makes it possible to customize the indention of if/case/recieve patterns.

Before the change, the indentaion of if/case/recieve patterns is
4 steps and it cannot be changed alone.  Some people prefer other
styles, for example, zero indentation.

case a of
true ->
   do_something();
false ->
   do_something_else()
end.

This patch just adds an option, which default value is 4 steps.
"(setq erlang-icr-indent 0)" will keep if/case/receive at the same column of
the sub-clauses.