Fix infinite loop for malformed edoc input
When processing an edoc comment with ``` in it, if the comment ends
without a matching ''' then an infinite loop occurs in the function
edoc_wiki:strip_empty_lines/2. This change fixes that by adding a clause
to return from the function upon the end of the comment input. This
allows an error to be thrown to indicate the problem, which is the same
behaviour as leaving either `` or ` unmatched.