Fix bug in the simplification of inexact comparisons
On 31/1/2011 Paul Guyot reported a bug in the native code compilation
of inexact equality/inequality tests between floats and integers. The
relevant test was:

f(X) ->
    Y = X / 2,
    Y == 0.

and hipe erroneously evaluated the calls f(0) and f(0.0) to 'false'.

The culprit was in the simplification code of the Icode range analysis
which used an erroneous test (lists:any/1 instead of lists:all/1).